+++ navigator/src/main/java/com/soa/navigator/model/Coordinates.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class Coordinates {
 
     @DecimalMin(value = "-868", message = "x should be greater thatn -868")
-    private Integer x;
+    private long x;
 
     @NotNull(message = "coordinates.x is not provided")
     @DecimalMin(value = "-355", message = "y should be greater thatn -355")
